Updated to version 4.3.1; changes see CHANGES.txt

+++ views/default/resources/izap_videos/videos/add.php
@@ -1,6 +1,6 @@
 <?php
 
-$guid = elgg_extract('guid', $vars);
+$guid = (int) elgg_extract('guid', $vars);
 if (!$guid) {
 	$guid = elgg_get_logged_in_user_guid();
 }
